To keep MSVC happy, clarify which overloaded version of 'floor()' to call

+++ gtk2_ardour/region_peak_cursor.cc
@@ -117,7 +117,7 @@ RegionPeakCursor::set (AudioRegionView* arv, samplepos_t when, samplecnt_t sampl
 
 	/* position relative to editor origin */
 	ArdourCanvas::Duple pos  = arv->get_canvas_group ()->item_to_window (_canvas_text->parent ()->position ());
-	double              xpos = pos.x + floor ((when - ar->position ()) / samples_per_pixel);
+	double              xpos = pos.x + floor ((double)(when - ar->position ()) / samples_per_pixel);
 
 	_canvas_text->set_x_position (xpos + 3);
 	_canvas_text->set_y_position (pos.y + 3);
